Added files via upload.
Project TARZAN is a device that aims to address the problem of illegal logging, hunting and human activity in the forests through smoke detection, PIR sensing and sound detection through an array of sensors and an Arduino connected to an XBEE RF module that can transmit data from node to node. Unfortunately, the Arduino code was not recovered.